+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+* Improve support for values sent in the POST body of a task request. These can now be included in
+  TaskRequestStub instances (provide the `parsed_body` option to ::with()). The TaskRequest also
+  includes a sugar method to grab a body (POST) value and throw if not present or empty.
+
 ## v0.2.2 (2021-04-09)
 
 * Fix MockTaskCreator so that it can compare the DateTimeImmutable for schedule_send_after and DateInterval for